requirements.txt
selenium
undetected-chromedriver
selenium-stealth
pandas
python-dotenv
webdriver-manager
setuptools